update workflow API
