+++ python/ray/experimental/workflow/tests/test_storage.py
@@ -19,7 +19,7 @@ def some_func2(x):
 
 @pytest.mark.asyncio
 async def test_raw_storage(workflow_start_regular):
-    raw_storage = storage.get_global_storage()
+    raw_storage = workflow_storage._StorageImpl(storage.get_global_storage())
     workflow_id = test_workflow_storage.__name__
     step_id = "some_step"
     input_metadata = {"2": "c"}
@@ -105,7 +105,7 @@ async def test_raw_storage(workflow_start_regular):
 
 
 def test_workflow_storage(workflow_start_regular):
-    raw_storage = storage.get_global_storage()
+    raw_storage = workflow_storage._StorageImpl(storage.get_global_storage())
     workflow_id = test_workflow_storage.__name__
     step_id = "some_step"
     input_metadata = {
@@ -144,7 +144,8 @@ def test_workflow_storage(workflow_start_regular):
                                               output_metadata))
     asyncio_run(raw_storage.save_step_output(workflow_id, step_id, output))
 
-    wf_storage = workflow_storage.WorkflowStorage(workflow_id, raw_storage)
+    wf_storage = workflow_storage.WorkflowStorage(workflow_id,
+                                                  storage.get_global_storage())
     assert wf_storage.load_step_output(step_id) == output
     assert wf_storage.load_step_args(step_id, [], [], []) == args
     assert wf_storage.load_step_func_body(step_id)(33) == 34
