+++ python/ray/experimental/workflow/workflow_storage.py
@@ -3,7 +3,9 @@ This module is higher-level abstraction of storage directly used by
 workflows.
 """
 
+import abc
 import asyncio
+import functools
 from typing import Dict, List, Optional, Any, Callable, Tuple, Union
 from dataclasses import dataclass
 
@@ -15,6 +17,22 @@ from ray.experimental.workflow.common import (Workflow, WorkflowData, StepID,
                                               WorkflowRef, StepType)
 from ray.experimental.workflow import workflow_context
 from ray.experimental.workflow import serialization_context
+from ray.experimental.workflow.storage import (DataLoadError, DataSaveError,
+                                               KeyNotFoundError)
+
+ArgsType = Tuple[List[Any], Dict[str, Any]]  # args and kwargs
+
+# constants used for keys
+OBJECTS_DIR = "objects"
+STEPS_DIR = "steps"
+STEP_INPUTS_METADATA = "inputs.json"
+STEP_OUTPUTS_METADATA = "outputs.json"
+STEP_ARGS = "args.pkl"
+STEP_OUTPUT = "output.pkl"
+STEP_FUNC_BODY = "func_body.pkl"
+CLASS_BODY = "class_body.pkl"
+WORKFLOW_META = "workflow_meta.json"
+WORKFLOW_PROGRESS = "progress.json"
 
 
 # TODO: Get rid of this and use asyncio.run instead once we don't support py36
@@ -56,12 +74,395 @@ class StepInspectResult:
                     and self.workflows is not None and self.func_body_valid))
 
 
+@dataclass
+class StepStatus:
+    # does the step output checkpoint exist?
+    output_object_exists: bool
+    # does the step output metadata exist?
+    output_metadata_exists: bool
+    # does the step input metadata exist?
+    input_metadata_exists: bool
+    # does the step input argument exist?
+    args_exists: bool
+    # does the step function body exist?
+    func_body_exists: bool
+
+
+def data_load_error(func):
+    @functools.wraps(func)
+    async def _func(*args, **kvargs):
+        try:
+            ret = await func(*args, **kvargs)
+            return ret
+        except Exception as e:
+            raise DataLoadError from e
+
+    return _func
+
+
+def data_save_error(func):
+    @functools.wraps(func)
+    async def _func(*args, **kv_args):
+        try:
+            ret = await func(*args, **kv_args)
+            return ret
+        except Exception as e:
+            raise DataSaveError from e
+
+    return _func
+
+
+class _StorageImpl(metaclass=abc.ABCMeta):
+    """Implementation for low level storage primitives, e.g. save the output
+    of a workflow step.
+    """
+
+    def __init__(self, store: "storage.Storage"):
+        self._store = store
+
+    async def _put(self, data: Any, paths: List[str],
+                   is_json: bool = False) -> None:
+        key = self._store.make_key(*paths)
+        await self._store.put(key, data, is_json=is_json)
+
+    async def _get(self, paths: List[str], is_json: bool = False) -> Any:
+        key = self._store.make_key(*paths)
+        return await self._store.get(key, is_json=is_json)
+
+    @data_load_error
+    async def load_step_input_metadata(self, workflow_id: str,
+                                       step_id: StepID) -> Dict[str, Any]:
+        """Load the input metadata of a step.
+
+        Args:
+            workflow_id: ID of the workflow job.
+            step_id: ID of the step.
+
+        Raises:
+            DataLoadError: if we fail to load the metadata.
+
+        Returns:
+            A metadata dict.
+        """
+        return await self._get(
+            [workflow_id, STEPS_DIR, step_id, STEP_INPUTS_METADATA], True)
+
+    @data_save_error
+    async def save_step_input_metadata(self, workflow_id: str, step_id: StepID,
+                                       metadata: Dict[str, Any]) -> None:
+        """Save the input metadata of a step.
+
+        Args:
+            workflow_id: ID of the workflow job.
+            step_id: ID of the step.
+            metadata: A metadata dict.
+
+        Raises:
+            DataSaveError: if we fail to save the metadata.
+        """
+        await self._put(
+            metadata, [workflow_id, STEPS_DIR, step_id, STEP_INPUTS_METADATA],
+            True)
+
+    @data_load_error
+    async def load_step_output_metadata(self, workflow_id: str,
+                                        step_id: StepID) -> Dict[str, Any]:
+        """Load the output metadata of a step.
+
+        Args:
+            workflow_id: ID of the workflow job.
+            step_id: ID of the step.
+
+        Raises:
+            DataLoadError: if we fail to load the metadata.
+
+        Returns:
+            A metadata dict.
+        """
+        return await self._get(
+            [workflow_id, STEPS_DIR, step_id, STEP_OUTPUTS_METADATA], True)
+
+    @data_save_error
+    async def save_step_output_metadata(self, workflow_id: str,
+                                        step_id: StepID,
+                                        metadata: Dict[str, Any]) -> None:
+        """Save the output metadata of a step.
+
+        Args:
+            workflow_id: ID of the workflow job.
+            step_id: ID of the step.
+            metadata: A metadata dict.
+
+        Raises:
+            DataSaveError: if we fail to save the metadata.
+        """
+        await self._put(
+            metadata, [workflow_id, STEPS_DIR, step_id, STEP_OUTPUTS_METADATA],
+            True)
+
+    @data_load_error
+    async def load_step_output(self, workflow_id: str, step_id: StepID) -> Any:
+        """Load the output of the workflow step from checkpoint.
+
+        Args:
+            workflow_id: ID of the workflow job.
+            step_id: ID of the workflow step.
+
+        Raises:
+            DataLoadError: if we fail to load the output.
+
+        Returns:
+            Output of the workflow step.
+        """
+        return await self._get([workflow_id, STEPS_DIR, step_id, STEP_OUTPUT])
+
+    @data_save_error
+    async def save_step_output(self, workflow_id: str, step_id: StepID,
+                               output: Any) -> None:
+        """Save the output of a workflow step.
+
+        Args:
+            workflow_id: ID of the workflow job.
+            output: The output object.
+
+        Raises:
+            DataSaveError: if we fail to save the output.
+        """
+        await self._put(output, [workflow_id, STEPS_DIR, step_id, STEP_OUTPUT])
+
+    @data_load_error
+    async def load_step_func_body(self, workflow_id: str,
+                                  step_id: StepID) -> Callable:
+        """Load the function body of the workflow step.
+
+        Args:
+            workflow_id: ID of the workflow job.
+            step_id: ID of the workflow step.
+
+        Raises:
+            DataLoadError: if we fail to load the function body.
+
+        Returns:
+            A callable function.
+        """
+        return await self._get(
+            [workflow_id, STEPS_DIR, step_id, STEP_FUNC_BODY])
+
+    @data_save_error
+    async def save_step_func_body(self, workflow_id: str, step_id: StepID,
+                                  func_body: Callable) -> None:
+        """Save the function body of the workflow step.
+
+        Args:
+            workflow_id: ID of the workflow job.
+            step_id: ID of the workflow step.
+            func_body: The step function to be written.
+
+        Raises:
+            DataSaveError: if we fail to save the function body.
+        """
+        await self._put(func_body,
+                        [workflow_id, STEPS_DIR, step_id, STEP_FUNC_BODY])
+
+    @data_load_error
+    async def load_step_args(self, workflow_id: str,
+                             step_id: StepID) -> ArgsType:
+        """Load the input arguments of the workflow step. This must be
+        done under a serialization context, otherwise the arguments would
+        not be reconstructed successfully.
+
+        Args:
+            workflow_id: ID of the workflow job.
+            step_id: ID of the workflow step.
+
+        Raises:
+            DataLoadError: if we fail to load the arguments.
+
+        Returns:
+            Args and kwargs.
+        """
+        return await self._get([workflow_id, STEPS_DIR, step_id, STEP_ARGS])
+
+    @data_save_error
+    async def save_step_args(self, workflow_id: str, step_id: StepID,
+                             args: ArgsType) -> None:
+        """Save the function body of the workflow step.
+
+        Args:
+            workflow_id: ID of the workflow job.
+            step_id: ID of the workflow step.
+            args: The step input args to be written.
+
+        Raises:
+            DataSaveError: if we fail to save the arguments.
+        """
+        await self._put(args, [workflow_id, STEPS_DIR, step_id, STEP_ARGS])
+
+    @data_load_error
+    async def load_object_ref(self, workflow_id: str,
+                              object_id: str) -> ray.ObjectRef:
+        """Load the input object ref.
+
+        Args:
+            workflow_id: ID of the workflow job.
+            object_id: The hex ObjectID.
+
+        Raises:
+            DataLoadError: if we fail to load the object ref.
+
+        Returns:
+            The object ref.
+        """
+        data = await self._get([workflow_id, OBJECTS_DIR, object_id])
+        return ray.put(data)  # simulate an ObjectRef
+
+    @data_save_error
+    async def save_object_ref(self, workflow_id: str,
+                              obj_ref: ray.ObjectRef) -> None:
+        """Save the input object ref.
+
+        Args:
+            workflow_id: ID of the workflow job.
+            obj_ref: The ObjectRef to be saved.
+
+        Raises:
+            DataSaveError: if we fail to save the object ref.
+        """
+        data = await obj_ref
+        await self._put(data, [workflow_id, OBJECTS_DIR, obj_ref.hex()])
+
+    @data_load_error
+    async def load_actor_class_body(self, workflow_id: str) -> type:
+        """Load the class body of the virtual actor.
+
+        Args:
+            workflow_id: ID of the workflow job.
+
+        Raises:
+            DataLoadError: if we fail to load the class body.
+        """
+        return await self._get([workflow_id, CLASS_BODY])
+
+    @data_save_error
+    async def save_actor_class_body(self, workflow_id: str, cls: type) -> None:
+        """Save the class body of the virtual actor.
+
+        Args:
+            workflow_id: ID of the workflow job.
+            cls: The class body used by the virtual actor.
+
+        Raises:
+            DataSaveError: if we fail to save the class body.
+        """
+        await self._put(cls, [workflow_id, CLASS_BODY])
+
+    @data_save_error
+    async def save_workflow_meta(self, workflow_id: str,
+                                 metadata: Dict[str, Any]) -> None:
+        """Save the meta of the workflow.
+
+        Args:
+            workflow_id: ID of the workflow
+            metadata: A metadata dict
+
+        Raises:
+            DataSaveError: if we fail to save the metadata.
+        """
+        await self._put(metadata, [workflow_id, WORKFLOW_META], True)
+
+    @data_load_error
+    async def load_workflow_meta(self,
+                                 workflow_id: str) -> Optional[Dict[str, Any]]:
+        """Load the meta of the workflow.
+
+        Args:
+            workflow_id: ID of the workflow
+
+        Raises:
+            DataLoadError: if we fail to load the metadata.
+
+        Returns:
+            A metadata dict or None.
+        """
+        try:
+            return await self._get([workflow_id, WORKFLOW_META], True)
+        except KeyNotFoundError:
+            return None
+
+    @data_load_error
+    async def load_workflow_progress(self, workflow_id: str) -> Dict[str, Any]:
+        """Load the latest progress of a workflow. This is used by a
+        virtual actor.
+
+        Args:
+            workflow_id: ID of the workflow job.
+
+        Raises:
+            DataLoadError: if we fail to load the progress.
+
+        Returns:
+            Metadata about the workflow progress.
+        """
+        return await self._get([workflow_id, STEPS_DIR, WORKFLOW_PROGRESS],
+                               True)
+
+    @data_save_error
+    async def save_workflow_progress(self, workflow_id: str,
+                                     metadata: Dict[str, Any]) -> None:
+        """Save the latest progress of a workflow. This is used by a
+        virtual actor.
+
+        Args:
+            workflow_id: ID of the workflow job.
+            metadata: Metadata about the workflow progress.
+
+        Raises:
+            DataSaveError: if we fail to save the progress.
+        """
+        await self._put(metadata, [workflow_id, STEPS_DIR, WORKFLOW_PROGRESS],
+                        True)
+
+    @data_load_error
+    async def get_step_status(self, workflow_id: str,
+                              step_id: StepID) -> StepStatus:
+        """Check the status of a step in the storage.
+
+        Args:
+            workflow_id: ID of the workflow job.
+            step_id: ID of the step.
+
+        Returns:
+            A dataclass of the step fields.
+        """
+        prefix = self._store.make_key(workflow_id, STEPS_DIR, step_id) + "/"
+        keys = set(await self._store.scan_prefix(prefix))
+        return StepStatus(
+            output_object_exists=(STEP_OUTPUT in keys),
+            output_metadata_exists=(STEP_OUTPUTS_METADATA in keys),
+            input_metadata_exists=(STEP_INPUTS_METADATA in keys),
+            args_exists=(STEP_ARGS in keys),
+            func_body_exists=(STEP_FUNC_BODY in keys))
+
+    @data_load_error
+    async def list_workflow(self) -> List[str]:
+        """List all the workflows inside the storage.
+
+        Raises:
+            DataLoadError: if we fail to load the metadata.
+
+        Returns:
+            A list of workflow ids
+        """
+        prefix = self._store.make_key("")
+        return await self._store.scan_prefix(prefix)
+
+
 class WorkflowStorage:
     """Access workflow in storage. This is a higher-level abstraction,
     which does not care about the underlining storage implementation."""
 
     def __init__(self, workflow_id: str, store: storage.Storage):
-        self._storage = store
+        self._storage = _StorageImpl(store)
         self._workflow_id = workflow_id
 
     def load_step_output(self, step_id: StepID) -> Any:
